Reduce route and CLI orchestrator blast radius
Split the oversized MCP route and agent-ask entrypoint into focused modules while preserving public entrypoints and server behavior. Refresh HTTP v2 harness wiring so the refactor remains covered against current auth and database initialization paths.

Constraint: Issue 635 requires mcp.routes.ts and agent-ask.ts orchestrators under 500 lines without MCP or HTTP v2 regressions.

Rejected: Leave test-http-server-v2 on stale internal imports | it no longer exercised the current server setup reliably.

Confidence: high

Scope-risk: moderate

Directive: Keep route and CLI entrypoints thin; add behavior to the extracted modules rather than rebuilding monoliths.

Tested: npm run build:packages; npm run test:http-v2; npm run test:server; npm run lint; npm run type-check; npm test; npm run check-debt-markers -- --production-only; graphify rebuild; git diff --check

Not-tested: production deployment smoke
